assets/scripts/download-assignment-submissions/requirements.txt
azure-core==1.20.1
azure-cosmos==4.2.0
azure-storage-blob==12.9.0
